+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* Add support for records with polymorphic fields.
 * Add the attribute `@coq_cast` to force an unsafe cast of a sub-expression.
 * Rename the attribute `@coq_axiom` as `@coq_axiom_with_reason`.
 * Add the attribute `@coq_precise_signature` to help to distinguish between ambiguous signatures.